Split ShapeTraversalRegistry from ModelTraversalRegistry with Shape specific logic
